+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project are documented in this file.
 ### Added
 
 - Added `blf querystring` and `blf qs` to parse query string params from a raw query string, full URL, or stdin via `-`.
+- Added `blf cal` to print previous, current, and next month calendars with week numbers.
 
 ## [v0.2.3] - 2026-04-04